+++ cmd/internal/utils_test.go
@@ -1,9 +1,10 @@
 package internal
 
 import (
-	"github.com/stretchr/testify/assert"
 	"strings"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestGetCwdFromExe(t *testing.T) {
@@ -12,7 +13,11 @@ func TestGetCwdFromExe(t *testing.T) {
 }
 
 func TestCheckPath(t *testing.T) {
-	assert.True(t, CheckPath("docker"), "Expected `CheckPath()` to find `docker` in `$PATH`")
+	dockerFound := CheckPath("docker")
+	if !dockerFound {
+		dockerFound = CheckPath("podman")
+	}
+	assert.True(t, dockerFound, "Expected `CheckPath()` to find `docker` or `podman` in `$PATH`")
 }
 
 func TestRunBasicCmd(t *testing.T) {